Don't parse route out of stop id
It's brittle and doesn't actually work well / have much meaning. It
seems to mostly be indicative of one of the routes that served a station
when it *first* opened (e.g. some stops near the end of the E in Queens
start w/ `G`, and that is far from unique). It also shows only one
route, making it completely useless; it doesn't even indicate which
GTFS feed is needed, because at several stations trains from multiple
divisions stop at the same GTFS stop (furthermore, the app continually
updates feeds in the background, not on-demand based on which station is
being requested).